+++ lib/super/controls.rb
@@ -1,25 +1,38 @@
 module Super
   class Controls
-    def initialize(dashboard)
-      @dashboard = dashboard
+    def initialize(actual)
+      @actual = actual
     end
 
-    attr_reader :dashboard
+    attr_reader :actual
 
-    def method_missing(method_name, *args)
-      if @dashboard.respond_to?(method_name)
-        @dashboard.public_send(method_name, *args)
-      else
-        super
-      end
+    def title
+      @actual.title
     end
 
-    def respond_to_missing?(method_name, _ = false)
-      if @dashboard.respond_to?(method_name)
-        true
-      else
-        super
-      end
+    def model
+      @actual.model
+    end
+
+    # @param action [ActionInquirer]
+    def scope(action:)
+      @actual.scope(action: action)
+    end
+
+    # @param params [ActionController::Parameters]
+    # @param action [ActionInquirer]
+    def permitted_params(params, action:)
+      @actual.permitted_params(params, action: action)
+    end
+
+    # @param action [ActionInquirer]
+    def display_schema(action:)
+      @actual.display_schema(action: action)
+    end
+
+    # @param action [ActionInquirer]
+    def form_schema(action:)
+      @actual.form_schema(action: action)
     end
   end
 end